Refactor S3Analytics
