+++ src/api/aws/index.js
@@ -1,5 +1,7 @@
 import * as Accounts from './accounts';
+import * as S3 from './s3';
 
 export default {
   Accounts,
+  S3
 };
