+++ src/api/index.js
@@ -1,9 +1,7 @@
 import * as Auth from './auth';
 import AWS from './aws';
-import * as s3 from './s3';
 
 export default {
   Auth,
   AWS,
-  s3,
 };
