+++ src/components/aws/index.js
@@ -1,5 +1,7 @@
 import Accounts from './accounts';
+import S3Analytics from './s3Analytics';
 
 export default {
-  Accounts
+  Accounts,
+  S3Analytics
 };
