+++ src/components/aws/s3Analytics/S3AnalyticsBarChartComponent.js
