+++ src/components/index.js
@@ -1,5 +1,4 @@
 import DeclareSetup from './DeclareSetupComponent';
-import S3Analytics from './s3Analytics';
 import Misc from './misc';
 import AWS from './aws';
 
@@ -7,5 +6,4 @@ export default {
   DeclareSetup,
   Misc,
   AWS,
-  S3Analytics,
 };
