+++ src/containers/aws/index.js
@@ -1,0 +1,5 @@
+import S3Analytics from "./s3AnalyticsContainer";
+
+export default {
+  S3Analytics
+};
