+++ src/containers/index.js
@@ -1,6 +1,5 @@
 import Main from './MainContainer';
-import Login from './auth/LoginContainer';
-import S3Analytics from './S3AnalyticsContainer';
+import AWS from './aws';
 import Auth from './auth';
 import Setup from './setup';
 
@@ -8,6 +7,5 @@ export default {
   Main,
   Auth,
   Setup,
-  Login,
-  S3Analytics,
+  AWS
 };
